// client/src/types.ts
export interface Message {
  _id?: string;
  sender: string;
  messageDateTime: Date;
  content: string;
  views: string[];
  likes: string[];
}

export interface Thread {
  _id?: string;
  messages: Message[];
  accounts: string[];
  threadUpdatedDateTime: Date;
}